% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_msaeOB.R
\name{est_msaeOB}
\alias{est_msaeOB}
\title{Ini Judul}
\usage{
est_msaeOB(
  formula,
  vardir,
  weight,
  samevar = FALSE,
  MAXITER = 100,
  PRECISION = 1e-04,
  data
)
}
\arguments{
\item{formula}{Penjelasan formula}

\item{vardir}{Penjelasan vardir}

\item{weight}{Penjelasan weight}

\item{samevar}{Penjelasan samevar}

\item{MAXITER}{Penjelasan MAXITER}

\item{PRECISION}{Penjelasan PRECISION}

\item{data}{Penjelasan data}
}
\value{
Ini return
}
\description{
Ini Judul
}
\examples{
Ini contoh
}
